can plot time dependent lesser and greater green's functions
The real part of the lesser and the real part of the greater (times negative 1) is considered, and the negative sign from the greater is ignored, since the negative sign was not taken into account when extracting the frequencies and weights in the angular frequency domain. It turns out that, if at half-filling, the real part of the lesser and the real part of the greater (times negative 1) are the same, so there is essentially one signal to pass to compressive sensing. And even if there are two distinct signals, one from lesser and one from greater, since the real part was chosen, a discrete cosine transform basis can be used for the compressive sensing, which is easier to use than a discrete sine transform basis. Both the lesser and greater green's function information is needed, just like in the exact case. Just choose real or imaginary, basically choosing cosine (even function) or sine (odd function).